DL-13479 split AmendsController (#521)
