+++ app/views/pages/amends/amendCurrentPensions.scala.html
@@ -66,7 +66,7 @@
         helpTextAccordionContent
     )
 
-    @form(action = routes.AmendsController.submitAmendCurrentPension) {
+    @form(action = routes.AmendsCurrentPensionController.submitAmendCurrentPension) {
 
         @formInputMoney(
             field = amendCurrentPensionForm("amendedUKPensionAmt"),
