+++ app/views/pages/amends/amendIP14OverseasPensions.scala.html
@@ -53,7 +53,7 @@
 
     <h1 class="govuk-heading-xl">@Messages("pla.overseasPensions.title")</h1>
 
-    @form(action = routes.AmendsController.submitAmendOverseasPensions) {
+    @form(action = routes.AmendsOverseasPensionController.submitAmendOverseasPensions) {
 
         <p class="govuk-body">@Messages("pla.ip14OverseasPensions.question")</p>
