+++ app/views/pages/amends/amendIP14PensionsTakenBefore.scala.html
@@ -83,7 +83,7 @@
 
     @errorSummary(amendPensionsTakenBeforeForm.errors, Some("amendedPensionsTakenBeforeAmt"))
 
-    @form(action = routes.AmendsController.submitAmendPensionsTakenBefore) {
+    @form(action = routes.AmendsPensionTakenBeforeController.submitAmendPensionsTakenBefore) {
 
         <h1 class="govuk-heading-xl">@Messages("pla.pensionsTakenBefore.title")</h1>
