+++ app/views/pages/amends/amendIP14PensionsUsedBetween.scala.html
@@ -66,7 +66,7 @@
     @errorSummary(amendPensionsUsedBetweenForm.errors, Some("amendedPensionsUsedBetweenAmt"))
     <h1 class="govuk-heading-xl">@Messages("pla.pensionsUsedBetween.question")</h1>
     <p class="govuk-body govuk-!-margin-top-4">@Messages("pla.pensionsUsedBetween.para.1")</p>
-    @form(action = routes.AmendsController.submitAmendPensionsUsedBetween) {
+    @form(action = routes.AmendsPensionUsedBetweenController.submitAmendPensionsUsedBetween) {
         @helpTextAccordion(
             id = "ip14-amend-pensions-used-between-help",
             summaryText = Messages("pla.pensionsUsedBetween.help"),
