+++ app/views/pages/amends/amendIP14PensionsWorthBefore.scala.html
@@ -70,7 +70,7 @@
 
      <h1 class="govuk-heading-xl">@Messages("pla.ip14PensionsTakenBefore.question")</h1>
 
-     @form(action = routes.AmendsController.submitAmendPensionsWorthBefore) {
+     @form(action = routes.AmendsPensionWorthBeforeController.submitAmendPensionsWorthBefore) {
           @helpTextAccordion(
                id = "ip14-amend-pensions-taken-before-help",
                summaryText = Messages("pla.pensionsWorthBefore.help"),
