+++ app/views/pages/amends/amendPensionsTakenBefore.scala.html
@@ -41,7 +41,7 @@
 
     <h1 class="govuk-heading-xl">@Messages("pla.pensionsTakenBefore.title")</h1>
 
-    @form(action = routes.AmendsController.submitAmendPensionsTakenBefore) {
+    @form(action = routes.AmendsPensionTakenBeforeController.submitAmendPensionsTakenBefore) {
 
         @plaInputRadioGroup(
             legend = Messages("pla.pensionsTakenBefore.title"),
