+++ app/views/pages/amends/amendPensionsTakenBetween.scala.html
@@ -45,7 +45,7 @@
 
     <h1 class="govuk-heading-l">@Messages("pla.pensionsTakenBetween.title")</h1>
 
-    @form(action = routes.AmendsController.submitAmendPensionsTakenBetween) {
+    @form(action = routes.AmendsPensionTakenBetweenController.submitAmendPensionsTakenBetween) {
 
     @plaInputRadioGroup(
         legend = Messages("pla.pensionsTakenBetween.title"),
