+++ app/views/pages/amends/amendPensionsWorthBefore.scala.html
@@ -70,7 +70,7 @@
 
      <h1 class="govuk-heading-xl">@Messages("pla.pensionsWorthBefore.title")</h1>
 
-     @form(action = routes.AmendsController.submitAmendPensionsWorthBefore) {
+     @form(action = routes.AmendsPensionWorthBeforeController.submitAmendPensionsWorthBefore) {
           @helpTextAccordion(
                id = "ip16-amend-pensions-taken-before-help",
                summaryText = Messages("pla.pensionsWorthBefore.help"),
