+++ app/views/pages/amends/removePsoDebits.scala.html
@@ -34,7 +34,7 @@
     <p class="govuk-body">@Messages("pla.amends.pensionDebits.removePso")</p>
 
 
-    @form(action = routes.AmendsController.submitRemovePso) {
+    @form(action = routes.AmendsRemovePensionSharingOrderController.submitRemovePso) {
 
         @submitButton(Messages("pla.base.remove"))
