+++ test/controllers/AmendsCurrentPensionControllerSpec.scala
@@ -1,0 +1,219 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers
+
+import auth.{AuthFunction, AuthFunctionImpl, authenticatedFakeRequest}
+import common.Exceptions.RequiredValueNotDefinedException
+import config._
+import config.wiring.PlaFormPartialRetriever
+import connectors.PLAConnector
+import constructors.{DisplayConstructors, ResponseConstructors}
+import enums.ApplicationType
+import forms.{AmendCurrentPensionForm, AmendOverseasPensionsForm, AmendPensionsTakenBeforeForm, AmendPensionsTakenBetweenForm}
+import mocks.AuthMock
+import models._
+import models.amendModels._
+import models.cache.CacheMap
+import org.apache.pekko.actor.ActorSystem
+import org.apache.pekko.stream.Materializer
+import org.jsoup.Jsoup
+import org.mockito.ArgumentMatchers.{any, anyString, startsWith}
+import org.mockito.Mockito._
+import org.scalatest.BeforeAndAfterEach
+import org.scalatestplus.mockito.MockitoSugar
+import play.api.Environment
+import play.api.http.HeaderNames.CACHE_CONTROL
+import play.api.i18n.{I18nSupport, Lang, Messages, MessagesApi}
+import play.api.libs.json.{JsNull, Json}
+import play.api.mvc.MessagesControllerComponents
+import play.api.test.FakeRequest
+import play.api.test.Helpers._
+import services.SessionCacheService
+import testHelpers._
+import uk.gov.hmrc.auth.core.retrieve.v2.Retrievals
+import uk.gov.hmrc.govukfrontend.views.html.components.FormWithCSRF
+import uk.gov.hmrc.http.HttpResponse
+import views.html.pages.amends._
+import views.html.pages.fallback.{noNotificationId, technicalError}
+import views.html.pages.result.manualCorrespondenceNeeded
+
+import java.time.LocalDate
+import java.util.UUID
+import scala.concurrent.{ExecutionContext, Future}
+
+class AmendsCurrentPensionControllerSpec extends FakeApplication
+  with MockitoSugar
+  with SessionCacheTestHelper
+  with BeforeAndAfterEach with AuthMock with I18nSupport {
+
+  implicit lazy val mockMessage = fakeApplication().injector.instanceOf[MessagesControllerComponents].messagesApi.preferred(fakeRequest)
+
+  val mockSessionCacheService: SessionCacheService       = mock[SessionCacheService]
+  val mockMCC: MessagesControllerComponents          = fakeApplication().injector.instanceOf[MessagesControllerComponents]
+  val mockAuthFunction: AuthFunction                 = mock[AuthFunction]
+  val mockTechnicalError: technicalError             = app.injector.instanceOf[technicalError]
+  val mockAmendCurrentPensions: amendCurrentPensions = app.injector.instanceOf[amendCurrentPensions]
+  val mockAmendIP14CurrentPensions: amendIP14CurrentPensions = app.injector.instanceOf[amendIP14CurrentPensions]
+  val mockEnv: Environment                            = mock[Environment]
+  val messagesApi: MessagesApi                        = mockMCC.messagesApi
+
+  implicit val partialRetriever: PlaFormPartialRetriever = mock[PlaFormPartialRetriever]
+  implicit val mockAppConfig: FrontendAppConfig = fakeApplication().injector.instanceOf[FrontendAppConfig]
+  implicit val mockPlaContext: PlaContext = mock[PlaContext]
+  implicit val system: ActorSystem = ActorSystem()
+  implicit val materializer: Materializer = mock[Materializer]
+  implicit val mockLang: Lang = mock[Lang]
+  implicit val formWithCSRF: FormWithCSRF = app.injector.instanceOf[FormWithCSRF]
+  implicit val ec: ExecutionContext = app.injector.instanceOf[ExecutionContext]
+
+  override def beforeEach() = {
+    reset(
+      mockSessionCacheService,
+      mockAuthConnector,
+      mockEnv
+    )
+    super.beforeEach()
+  }
+
+  val testIP16DormantModel = AmendProtectionModel(ProtectionModel(None, None), ProtectionModel(None, None, protectionType = Some("IP2016"), status = Some("dormant"), relevantAmount = Some(100000), uncrystallisedRights = Some(100000)))
+
+  class Setup {
+    val authFunction = new AuthFunctionImpl (
+      mockMCC,
+      mockAuthConnector,
+      mockTechnicalError
+      )
+
+    val controller = new AmendsCurrentPensionController(
+      mockSessionCacheService,
+      mockMCC,
+      authFunction,
+      mockTechnicalError,
+      mockAmendCurrentPensions,
+      mockAmendIP14CurrentPensions
+    )
+  }
+
+  implicit val fakeRequest = FakeRequest()
+
+  val ip2014Protection = ProtectionModel(
+    psaCheckReference = Some("testPSARef"),
+    uncrystallisedRights = Some(100000.00),
+    nonUKRights = Some(2000.00),
+    preADayPensionInPayment = Some(2000.00),
+    postADayBenefitCrystallisationEvents = Some(2000.00),
+    notificationId = Some(12),
+    protectionID = Some(12345),
+    protectionType = Some("IP2014"),
+    status = Some("dormant"),
+    certificateDate = Some("2016-04-17"),
+    protectedAmount = Some(1250000),
+    protectionReference = Some("PSA123456"))
+
+  val testAmendIP2014ProtectionModel = AmendProtectionModel(ip2014Protection, ip2014Protection)
+
+  def cacheFetchCondition[T](data: Option[T]): Unit = {
+    when(mockSessionCacheService.fetchAndGetFormData[T](anyString())(any(), any()))
+      .thenReturn(Future.successful(data))
+  }
+
+
+  "Calling the .amendCurrentPensions action" when {
+
+    "not supplied with a stored model" in new Setup {
+      mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+      cacheFetchCondition[AmendProtectionModel](None)
+
+      val result = controller.amendCurrentPensions("ip2016", "open")(fakeRequest)
+      status(result) shouldBe 500
+    }
+
+    "supplied with a stored test model (£100000, IP2016, dormant)" in new Setup {
+      val testModel = new AmendProtectionModel(ProtectionModel(None, None), ProtectionModel(None, None, uncrystallisedRights = Some(100000)))
+      lazy val result = controller.amendCurrentPensions("ip2016", "dormant")(fakeRequest)
+      lazy val jsoupDoc = Jsoup.parse(contentAsString(result))
+
+      mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+      cacheFetchCondition[AmendProtectionModel](Some(testModel))
+      status(result) shouldBe 200
+
+      cacheFetchCondition[AmendProtectionModel](Some(testModel))
+      jsoupDoc.body.getElementsByTag("h1").text shouldEqual Messages("pla.currentPensions.title")
+    }
+
+      "return some HTML that" should {
+
+        "contain some text and use the character set utf-8" in new Setup {
+          val testModel = new AmendProtectionModel(ProtectionModel(None, None), ProtectionModel(None, None, uncrystallisedRights = Some(100000)))
+          lazy val result = controller.amendCurrentPensions("ip2016", "dormant")(fakeRequest)
+          mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+          cacheFetchCondition[AmendProtectionModel](Some(testModel))
+
+          contentType(result) shouldBe Some("text/html")
+          charset(result) shouldBe Some("utf-8")
+        }
+
+        "have the value 100000 completed in the amount input by default" in new Setup {
+          lazy val result = controller.amendCurrentPensions("ip2016", "dormant")(fakeRequest)
+          lazy val jsoupDoc = Jsoup.parse(contentAsString(result))
+          val testModel = new AmendProtectionModel(ProtectionModel(None, None), ProtectionModel(None, None, uncrystallisedRights = Some(100000)))
+
+          mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+          cacheFetchCondition[AmendProtectionModel](Some(testModel))
+
+          jsoupDoc.body.getElementById("amendedUKPensionAmt").attr("value") shouldBe "100000"
+        }
+      }
+    }
+
+    "supplied with a stored test model (£100000, IP2014, dormant)" in new Setup {
+      lazy val result = controller.amendCurrentPensions("ip2016", "dormant")(fakeRequest)
+        mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+        cacheFetchCondition[AmendProtectionModel](Some(testAmendIP2014ProtectionModel))
+
+        status(result) shouldBe 200
+    }
+
+
+  "Submitting Amend IP16 Current Pensions data" when {
+
+    "the data is valid" in new Setup {
+      object DataItem extends AuthorisedFakeRequestToPost(controller.submitAmendCurrentPension, ("amendedUKPensionAmt", "100000"), ("protectionType", "ip2016"), ("status", "dormant"))
+
+        mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+        cacheSaveCondition[AmendProtectionModel](mockSessionCacheService)
+        cacheFetchCondition[AmendProtectionModel](Some(testIP16DormantModel))
+
+        status(DataItem.result) shouldBe 303
+        redirectLocation(DataItem.result) shouldBe Some(s"${routes.AmendsController.amendsSummary("ip2016", "dormant")}")
+    }
+
+    "the data is invalid" in new Setup {
+      object DataItem extends AuthorisedFakeRequestToPost(controller.submitAmendCurrentPension, ("amendedUKPensionAmt", ""))
+        mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+        status(DataItem.result) shouldBe 400
+      }
+
+    "the model can't be fetched from cache" in new Setup {
+      object DataItem extends AuthorisedFakeRequestToPost(controller.submitAmendCurrentPension, ("amendedUKPensionAmt", "1000000"), ("protectionType", "IP2016"), ("status", "dormant"))
+        mockAuthRetrieval[Option[String]](Retrievals.nino, Some("AB123456A"))
+        cacheFetchCondition[AmendProtectionModel](None)
+
+        status(DataItem.result) shouldBe 500
+      }
+    }
+}
